[16.0][ADD] purchase_order_contract_kmitl
