+++ purchase_order_contract_kmitl/__manifest__.py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
 {
-    'name': 'Purchase Order Contract Name',
+    'name': 'Purchase Order Contract KMITL',
     'version': '16.0.1.0.0',
-    'summary': """ Purchase Order Contract Name Summary """,
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
     "category": "KMITL",
     'depends': ['purchase'],
     "data": [
-        "views/purchase_order_views.xml"
+        "views/purchase_order_views.xml",
+        "data/purchase_exception.xml"
     ],
     'installable': True,
     'auto_install': False,
