+++ purchase_order_contract_kmitl/data/purchase_exception.xml
@@ -1,0 +1,12 @@
+<odoo>
+    <record id="purchase_order_contract_exception" model="exception.rule">
+        <field name="name">ต้องระบุชื่อสัญญาและเลขที่สัญญา</field>
+        <field name="description">ข้อมูลใบสั่งซื้อ/จ้าง</field>
+        <field name="model">purchase.order</field>
+        <field name="exception_type">by_py_code</field>
+        <field name="code">
+if not self.contract_name or not self.contract_number:
+    failed = True
+        </field>
+    </record>
+</odoo>
