+++ purchase_order_contract_kmitl/views/purchase_order_views.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="view_purchase_order_form_contract" model="ir.ui.view">
+        <field name="name">view.purchase.order.form (Contract KMITL)</field>
+        <field name="model">purchase.order</field>
+        <field name="inherit_id" ref="purchase.purchase_order_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//group/group[2]" position="inside">
+                <field name="contract_name" />
+                <field name="contract_number" />
+            </xpath>
+        </field>
+    </record>
+</odoo>
