--- purchase_order_contract_name/models/purchase_order.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-from odoo import _, api, fields, models
-from odoo.exceptions import UserError, ValidationError
-
-
-class PurchaseOrder(models.Model):
-    _inherit = 'purchase.order'
-
-    READONLY_STATES = {
-        'purchase': [('readonly', True)],
-        'done': [('readonly', True)],
-        'cancel': [('readonly', True)],
-    }
-
-    contract_name = fields.Char(
-        string="contract name",
-        tracking=True,
-        states=READONLY_STATES,
-    )
