--- purchase_order_contract_name/views/purchase_order_views.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-  <odoo>
-    <record id="view_purchase_order_form" model="ir.ui.view">
-        <field name="name">view.purchase.order.form</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase.purchase_order_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//div[@name='reminder']/.." position="inside">
-                <field name="contract_name" required="1"/>
-            </xpath>
-        </field>
-    </record>
-  </odoo>
